crawler-service/requirements.txt
PyYAML==6.0.2
aiohttp==3.9.5
feedparser==6.0.11
kafka-python==2.0.2
python-dotenv==1.0.1
